Add URL reduction feature.

+++ app/javascript/components/Reduce.js
@@ -113,7 +113,7 @@ class Reduce extends React.Component {
                     <div className="pt-2 pb-2 pl-1 bg-light">
                     {(() => {
                       if('long' in this.state.reduced_url.errors){
-                        return this.state.reduced_url.errors.short[0];
+                        return this.state.reduced_url.errors.long[0];
                       }
                       else {
                         return 'Something went wrong, try again.';
